usecases modified and upload file feature added

+++ src/worker/application/delete_movie_usecase.py
@@ -1,7 +1,9 @@
 from shared.domain import Response, SuccessResponse
 from shared.infrastructure import ErrorResponse
 from ..domain import MongoRepository
+import pathlib
 import uuid
+import os
 
 
 class DeleteMovieUseCase:
@@ -13,16 +15,27 @@ class DeleteMovieUseCase:
         _movie = self._mongo_service.get_movie_by_id(id_movie)
         if _movie is None:
             raise ErrorResponse(
-                "Movie does not exist. Try again",
+                "Failed to get movie. Cannot verify if the movie exists",
                 self.transaction_id,
                 500
             )
+        if not _movie:
+            raise ErrorResponse(
+                "Movie does not exist. Try again with different ID",
+                self.transaction_id,
+                404
+            )
         result = self._mongo_service.delete_movie(id_movie)
-        if result is None:
+        if not result:
             raise ErrorResponse(
                 "Failed to delete movie. Try again",
                 self.transaction_id,
                 500
             )
+        try:
+            path = f"{pathlib.Path(__file__).cwd()}/src/{_movie.get('url', '')}"
+            os.remove(path)
+        except Exception:
+            pass
 
         return SuccessResponse({"status": "Movie deleted with success"}, 200, self.transaction_id)
